Add editing watched posts functionality
Implements one of the feature requests from #1.

This changes the action of the 'watch' button that is added under posts.
If a post has already been watched:
 - clicking this button will now show the watch dialog instead of
   automatically ending the watch like it used to
 - the watch dialog will now have an additional 'stop watching' button
   and the text for the 'watch post' button now reads 'save changes'

This approach allows users to easily change the watch options directly
from the post without having to clutter the main extension dialog with
another button such as 'edit'.